feat: Add admin page for Student Council.
Student Council requires to keep track of which rooms are occupied by
which club so that they can control the room bookings. This feature
helps achieve the Council to get information regarding which event is
going to be held at a particular room at a particular day.

feat: Additional feature requests by SC on View Bookings (#69)

* Add error message if no event is found

* Allow Operations to access View Bookings page

Change name from ViewBookings to ViewBookings component