fix: Critical bug fix involving the room booking logic
This fixes the mysterious anonymous rooms in roomsx error.
For eg. Consider there is a booking for the 29th. Now when someone
books rooms from 28-30 the rooms for 29th will incorrectly be added to
the roomsx array for 28 and 30 also. This was caused because taken rooms
was calculated across all booking dates and then when roomsx was to be
updated the same array of takenRooms concatenated with selectedRooms was
used to update all dates.